#62 refactor: 노후준비종합진단 폴더 변경
